Automated Report Generation/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Evaluation metrics
evaluate>=0.4.0
bert-score>=0.3.13

# NLP tools
spacy>=3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
